requirements.txt
torch>=2.2
torchvision>=0.17
transformers==4.44.2
accelerate>=0.33.0
bitsandbytes>=0.43.0
xformers>=0.0.27
sentencepiece
safetensors
pillow
numpy
pandas
tqdm
pyyaml
